+++ test/test_info.ml
@@ -7,7 +7,7 @@ let%expect_test "info" =
       let%bind link =
         Connection.call_exn
           connection
-          (Api.info (Id [ `Link (Thing.Link.Id.of_string "hmjd8r") ]) ())
+          (Endpoint.info (Id [ `Link (Thing.Link.Id.of_string "hmjd8r") ]) ())
         >>| List.hd_exn
       in
       let link =
@@ -35,7 +35,7 @@ let%expect_test "info__by_subreddit_name" =
       let%bind subreddits =
         Connection.call_exn
           connection
-          (Api.info
+          (Endpoint.info
              (Subreddit_name
                 (List.map [ "ocaml"; "redditdev"; "python" ] ~f:Subreddit_name.of_string))
              ())