+++ test/test_listings.ml
@@ -4,7 +4,7 @@ open! Import
 
 let%expect_test "best" =
   with_cassette "best" ~f:(fun connection ->
-      let%bind listing = Connection.call_exn connection (Api.best ()) in
+      let%bind listing = Connection.call_exn connection (Endpoint.best ()) in
       let child_ids = Listing.children listing |> List.map ~f:Thing.Link.id in
       print_s [%sexp (child_ids : Thing.Link.Id.t list)];
       [%expect
@@ -18,7 +18,9 @@ let%expect_test "best" =
 let%expect_test "links_by_id" =
   with_cassette "links_by_id" ~f:(fun connection ->
       let links = List.map ~f:Thing.Link.Id.of_string [ "icqrut"; "ikksvt" ] in
-      let%bind listing = Connection.call_exn connection (Api.links_by_id () ~links) in
+      let%bind listing =
+        Connection.call_exn connection (Endpoint.links_by_id () ~links)
+      in
       let ids = Listing.children listing |> List.map ~f:Thing.Link.id in
       print_s [%sexp (ids : Thing.Link.Id.t list)];
       [%expect {| (icqrut ikksvt) |}];
@@ -30,7 +32,7 @@ let%expect_test "random" =
       let%bind link_id =
         Connection.call_exn
           connection
-          (Api.random ~subreddit:(Subreddit_name.of_string "ocaml") ())
+          (Endpoint.random ~subreddit:(Subreddit_name.of_string "ocaml") ())
       in
       print_s [%sexp (link_id : Thing.Link.Id.t)];
       [%expect {| feyhbv |}];